fixed weather app
