+++ my-app/src/app/app-routing.module.ts
@@ -1,11 +1,16 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+import { CityListComponent } from './city-list/city-list.component';
 import { MainComponent } from './main/main.component';
 import { NotFoundPageComponent } from './not-found-page/not-found-page.component';
 import { ShowWeatherComponent } from './show-weather/show-weather.component';
 
 const routes: Routes = [
   { path: 'showweather', component: ShowWeatherComponent },
+  {
+    path: 'citylist',
+    component: CityListComponent,
+  },
   {
     path: 'main',
     component: MainComponent,
