+++ my-app/src/app/app.module.ts
@@ -6,9 +6,10 @@ import { AppComponent } from './app.component';
 import { ShowWeatherComponent } from './show-weather/show-weather.component';
 import { MainComponent } from './main/main.component';
 import { NotFoundPageComponent } from './not-found-page/not-found-page.component';
+import { CityListComponent } from './city-list/city-list.component';
 
 @NgModule({
-  declarations: [AppComponent, ShowWeatherComponent, MainComponent, NotFoundPageComponent],
+  declarations: [AppComponent, ShowWeatherComponent, MainComponent, NotFoundPageComponent, CityListComponent],
   imports: [BrowserModule, HttpClientModule, AppRoutingModule],
   providers: [],
   bootstrap: [AppComponent],
