+++ my-app/src/app/main/main.component.html
@@ -1,12 +1,5 @@
 <div class="container">
-  <ul>
-    <li *ngFor="let city of cities">
-      <button class="btnShowWeather" (click)="showWeather(city)">
-        {{ city }}
-      </button>
-    </li>
-    <button class="btnShowCities" (click)="showCities()">
-      Show all cities in California
-    </button>
-  </ul>
+  <button class="btnShowCities" (click)="showCities()">
+    Show all cities in California
+  </button>
 </div>
