+++ my-app/src/app/show-weather/show-weather.component.html
@@ -1,7 +1,7 @@
 <div class="links">
   <nav>
     <a routerLink="">Main</a>
-    <a routerLink="/showweather">Back</a>
+    <a routerLink="/citylist">Back</a>
   </nav>
 </div>
 <div class="weather">
